requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
twilio==9.0.4
python-dotenv==1.0.1
